+++ CHANGELOG.md
@@ -76,6 +76,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [#737](http://github.com/Nelson-numerical-software/nelson/issues/737): profiler failed with `sind` example.
 
+- [#739](http://github.com/Nelson-numerical-software/nelson/issues/739): special case with empty cell.
+
 - [#741](http://github.com/Nelson-numerical-software/nelson/issues/741): `ones` and `zeros` do not manager logical as typename.
 
 ## 0.6.8 (2022-08-27)